Use lookup instead of resolve in ns_pool_for_referral glue loop
This eliminates a source of infinite recursion and recursion amplification for glue-less NS
referrals.  In the previous code, there were two problems: the current lookup depth wasn't
passed to resolve, so there was the potential for infinite recursion when trying to resolve
glue records, since resolve could be made to re-invoke ns_pool_for_referral with another
NS referral. Second, resolve was called in a for loop over record types A and AAAA, so each
iteration could double the number of calls to resolve.

So, instead, we call ns_pool_for_zone for the NS referral names and then lookup on the returned
pool.  You can still make this code path re-invoke ns_pool_for_referral, but it is now subject
to the running depth limit, and the lookup amplification won't occur since the only call inside
the A,AAAA loop is lookup.